Fix integer literals parsing

+++ engine_test.go
@@ -582,7 +582,7 @@ var queries = []struct {
 	{
 		`SELECT COALESCE(NULL, NULL, NULL, COALESCE(NULL, 1234567890))`,
 		[]sql.Row{
-			{int64(1234567890)},
+			{int32(1234567890)},
 		},
 	},
 	{
@@ -981,7 +981,7 @@ var queries = []struct {
 	},
 	{
 		`SELECT -1`,
-		[]sql.Row{{int64(-1)}},
+		[]sql.Row{{int8(-1)}},
 	},
 	{
 		`
@@ -1043,13 +1043,13 @@ var queries = []struct {
 	{
 		`SELECT nullif(123, 321)`,
 		[]sql.Row{
-			{int64(123)},
+			{int8(123)},
 		},
 	},
 	{
 		`SELECT ifnull(123, NULL)`,
 		[]sql.Row{
-			{int64(123)},
+			{int8(123)},
 		},
 	},
 	{
@@ -1061,19 +1061,19 @@ var queries = []struct {
 	{
 		`SELECT ifnull(NULL, 123)`,
 		[]sql.Row{
-			{int64(123)},
+			{int8(123)},
 		},
 	},
 	{
 		`SELECT ifnull(123, 123)`,
 		[]sql.Row{
-			{int64(123)},
+			{int8(123)},
 		},
 	},
 	{
 		`SELECT ifnull(123, 321)`,
 		[]sql.Row{
-			{int64(123)},
+			{int8(123)},
 		},
 	},
 	{
@@ -1085,7 +1085,7 @@ var queries = []struct {
 	{
 		`SELECT round(15, 1)`,
 		[]sql.Row{
-			{int64(15)},
+			{int8(15)},
 		},
 	},
 	{
@@ -1452,7 +1452,7 @@ var queries = []struct {
 	},
 	{
 		`SELECT 1 FROM mytable GROUP BY i HAVING i > 1`,
-		[]sql.Row{{int64(1)}, {int64(1)}},
+		[]sql.Row{{int8(1)}, {int8(1)}},
 	},
 	{
 		`SELECT avg(i) FROM mytable GROUP BY i HAVING avg(i) > 1`,
@@ -1887,8 +1887,8 @@ func TestInsertInto(t *testing.T) {
 			[]sql.Row{{int64(1)}},
 			"SELECT * FROM typestable WHERE id = 999;",
 			[]sql.Row{{
-				int64(999), int64(math.MaxInt8), int64(math.MaxInt16), int64(math.MaxInt32), int64(math.MaxInt64),
-				int64(math.MaxUint8), int64(math.MaxUint16), int64(math.MaxUint32), uint64(math.MaxUint64),
+				int64(999), int8(math.MaxInt8), int16(math.MaxInt16), int32(math.MaxInt32), int64(math.MaxInt64),
+				uint8(math.MaxUint8), uint16(math.MaxUint16), uint32(math.MaxUint32), uint64(math.MaxUint64),
 				float64(math.MaxFloat32), float64(math.MaxFloat64),
 				timeParse(sql.TimestampLayout, "2132-04-05 12:51:36"), timeParse(sql.DateLayout, "2231-11-07"),
 				"random text", true, `{"key":"value"}`, "blobdata",
@@ -1905,8 +1905,8 @@ func TestInsertInto(t *testing.T) {
 			[]sql.Row{{int64(1)}},
 			"SELECT * FROM typestable WHERE id = 999;",
 			[]sql.Row{{
-				int64(999), int64(math.MaxInt8), int64(math.MaxInt16), int64(math.MaxInt32), int64(math.MaxInt64),
-				int64(math.MaxUint8), int64(math.MaxUint16), int64(math.MaxUint32), uint64(math.MaxUint64),
+				int64(999), int8(math.MaxInt8), int16(math.MaxInt16), int32(math.MaxInt32), int64(math.MaxInt64),
+				uint8(math.MaxUint8), uint16(math.MaxUint16), uint32(math.MaxUint32), uint64(math.MaxUint64),
 				float64(math.MaxFloat32), float64(math.MaxFloat64),
 				timeParse(sql.TimestampLayout, "2132-04-05 12:51:36"), timeParse(sql.DateLayout, "2231-11-07"),
 				"random text", true, `{"key":"value"}`, "blobdata",
@@ -1923,8 +1923,8 @@ func TestInsertInto(t *testing.T) {
 			[]sql.Row{{int64(1)}},
 			"SELECT * FROM typestable WHERE id = 999;",
 			[]sql.Row{{
-				int64(999), int64(-math.MaxInt8 - 1), int64(-math.MaxInt16 - 1), int64(-math.MaxInt32 - 1), int64(-math.MaxInt64 - 1),
-				int64(0), int64(0), int64(0), int64(0),
+				int64(999), int8(-math.MaxInt8 - 1), int16(-math.MaxInt16 - 1), int32(-math.MaxInt32 - 1), int64(-math.MaxInt64 - 1),
+				uint8(0), uint16(0), uint32(0), uint64(0),
 				float64(math.SmallestNonzeroFloat32), float64(math.SmallestNonzeroFloat64),
 				timeParse(sql.TimestampLayout, "0010-04-05 12:51:36"), timeParse(sql.DateLayout, "0101-11-07"),
 				"", false, ``, "",
@@ -1941,8 +1941,8 @@ func TestInsertInto(t *testing.T) {
 			[]sql.Row{{int64(1)}},
 			"SELECT * FROM typestable WHERE id = 999;",
 			[]sql.Row{{
-				int64(999), int64(-math.MaxInt8 - 1), int64(-math.MaxInt16 - 1), int64(-math.MaxInt32 - 1), int64(-math.MaxInt64 - 1),
-				int64(0), int64(0), int64(0), int64(0),
+				int64(999), int8(-math.MaxInt8 - 1), int16(-math.MaxInt16 - 1), int32(-math.MaxInt32 - 1), int64(-math.MaxInt64 - 1),
+				uint8(0), uint16(0), uint32(0), uint64(0),
 				float64(math.SmallestNonzeroFloat32), float64(math.SmallestNonzeroFloat64),
 				timeParse(sql.TimestampLayout, "0010-04-05 12:51:36"), timeParse(sql.DateLayout, "0101-11-07"),
 				"", false, ``, "",
@@ -2087,8 +2087,8 @@ func TestReplaceInto(t *testing.T) {
 			[]sql.Row{{int64(1)}},
 			"SELECT * FROM typestable WHERE id = 999;",
 			[]sql.Row{{
-				int64(999), int64(math.MaxInt8), int64(math.MaxInt16), int64(math.MaxInt32), int64(math.MaxInt64),
-				int64(math.MaxUint8), int64(math.MaxUint16), int64(math.MaxUint32), uint64(math.MaxUint64),
+				int64(999), int8(math.MaxInt8), int16(math.MaxInt16), int32(math.MaxInt32), int64(math.MaxInt64),
+				uint8(math.MaxUint8), uint16(math.MaxUint16), uint32(math.MaxUint32), uint64(math.MaxUint64),
 				float64(math.MaxFloat32), float64(math.MaxFloat64),
 				timeParse(sql.TimestampLayout, "2132-04-05 12:51:36"), timeParse(sql.DateLayout, "2231-11-07"),
 				"random text", true, `{"key":"value"}`, "blobdata",
@@ -2105,8 +2105,8 @@ func TestReplaceInto(t *testing.T) {
 			[]sql.Row{{int64(1)}},
 			"SELECT * FROM typestable WHERE id = 999;",
 			[]sql.Row{{
-				int64(999), int64(math.MaxInt8), int64(math.MaxInt16), int64(math.MaxInt32), int64(math.MaxInt64),
-				int64(math.MaxUint8), int64(math.MaxUint16), int64(math.MaxUint32), uint64(math.MaxUint64),
+				int64(999), int8(math.MaxInt8), int16(math.MaxInt16), int32(math.MaxInt32), int64(math.MaxInt64),
+				uint8(math.MaxUint8), uint16(math.MaxUint16), uint32(math.MaxUint32), uint64(math.MaxUint64),
 				float64(math.MaxFloat32), float64(math.MaxFloat64),
 				timeParse(sql.TimestampLayout, "2132-04-05 12:51:36"), timeParse(sql.DateLayout, "2231-11-07"),
 				"random text", true, `{"key":"value"}`, "blobdata",
@@ -2123,8 +2123,8 @@ func TestReplaceInto(t *testing.T) {
 			[]sql.Row{{int64(1)}},
 			"SELECT * FROM typestable WHERE id = 999;",
 			[]sql.Row{{
-				int64(999), int64(-math.MaxInt8 - 1), int64(-math.MaxInt16 - 1), int64(-math.MaxInt32 - 1), int64(-math.MaxInt64 - 1),
-				int64(0), int64(0), int64(0), int64(0),
+				int64(999), int8(-math.MaxInt8 - 1), int16(-math.MaxInt16 - 1), int32(-math.MaxInt32 - 1), int64(-math.MaxInt64 - 1),
+				uint8(0), uint16(0), uint32(0), uint64(0),
 				float64(math.SmallestNonzeroFloat32), float64(math.SmallestNonzeroFloat64),
 				timeParse(sql.TimestampLayout, "0010-04-05 12:51:36"), timeParse(sql.DateLayout, "0101-11-07"),
 				"", false, ``, "",
@@ -2141,8 +2141,8 @@ func TestReplaceInto(t *testing.T) {
 			[]sql.Row{{int64(1)}},
 			"SELECT * FROM typestable WHERE id = 999;",
 			[]sql.Row{{
-				int64(999), int64(-math.MaxInt8 - 1), int64(-math.MaxInt16 - 1), int64(-math.MaxInt32 - 1), int64(-math.MaxInt64 - 1),
-				int64(0), int64(0), int64(0), int64(0),
+				int64(999), int8(-math.MaxInt8 - 1), int16(-math.MaxInt16 - 1), int32(-math.MaxInt32 - 1), int64(-math.MaxInt64 - 1),
+				uint8(0), uint16(0), uint32(0), uint64(0),
 				float64(math.SmallestNonzeroFloat32), float64(math.SmallestNonzeroFloat64),
 				timeParse(sql.TimestampLayout, "0010-04-05 12:51:36"), timeParse(sql.DateLayout, "0101-11-07"),
 				"", false, ``, "",
@@ -2947,7 +2947,7 @@ func TestSessionVariables(t *testing.T) {
 	rows, err := sql.RowIterToRows(iter)
 	require.NoError(err)
 
-	require.Equal([]sql.Row{{int64(1), ",STRICT_TRANS_TABLES"}}, rows)
+	require.Equal([]sql.Row{{int8(1), ",STRICT_TRANS_TABLES"}}, rows)
 }
 
 func TestSessionVariablesONOFF(t *testing.T) {
